refactor(review): decompose runSemanticReview into named stages (TYPE-1)
461-line runSemanticReview split into a thin orchestrator plus named
helpers in a new semantic-outcomes.ts sibling (dispatch, finding
classification, outcome builders), mirroring the adversarial.ts split.
Control flow, logging, and audit-recording order are preserved exactly —
only the grouping changed. Keeps both files under the 600-line limit.
